Create SliceRefinements.dfy
An important part of the framework for code-motion refactoring, with implementation of the proof method of slice refinements, based on Chapter 5 ("Proof Method for Correct Slicing-Based Refactoring") of my doctoral thesis ("Refactoring via Program Slicing and Sliding").

The initially uploaded version is an initial (partial) implementation of "A Theoretical Framework" (Chapter 4 of my thesis). This implementation comprises a formulation of the weakest-preconditions semantics of a simple program notation (a deterministic subset of Dijkstra's guarded commands + support for the declaration of local variables); it also includes an example transformation (5.7, from Chapter 5 of my thesis) written here in Dafny by my graduate student Or Golov; and a first attempt at formulating the requirements RE1-RE5 (from Chapter 4) on each language construct, as lemmas in Dafny, written with partial (incomplete) proofs by Or Golov.