Formulate Law 7
Introduction/removal of an assertion following an assignment.
For now this is formulated for assignments in which the order of the LHS
is conveniently a concatenation of irrelevant variables, X, and relevant
ones (for the assertion), Y. Any interleaving of such X and Y should be
allowed. Not sure how to implement this, though: it could be that the
representation of assignments should be modified to be a map from
variable name to a corresponding expression.