Formulate Law 18 parts (a), (b), and (c)
After an assignment (to Y) of expressions with non-assigned variables
(E2), all occurrences of Y can be replaced by the corresponding
expressions  in E2 in the following statement: on the RHS of an
assignment, on the predicate of a conditional, and on the predicate of
an iteration. The latter is correct only when the equality of Y and E2
is preserved by the loop body.