Upgrade to Dafny 2.0.0
Update function types to the new long arrow syntax: regular arrows
("->") not seem to stand for total functions, not the partial functions
that they used to denote.  The new syntax for partial functions is
through either "-->" for functions that may not read from the heap and
"~>" for those that may. We go for "-->" as a first step; yet in the
future we may want to benefit from Dafny's total functions, following
the Dijkstra and Scholten assumption(from their 1990 Monograph
"Predicate Calculus and Program Semantics").

Having modified all functions to "-->", all "reads" clauses were
removed. (They were anyway redundant: the vast majority of these clauses
were the unhelpful "reads *", that had been introduced merely to silence
Dafny's concerns.)

In addition, "this" can no longer be referred to in a precondition to a
constructor - so for now the postcondition of the VariablesSSA
constructor is commented out, as it is otherwise not guaranteed to hold,
since the forced removal of the (proably incorrect)  precondition that
it held on entry to the constructor.