Fix proof of Corollary_5_2HelpToCorollary_5_4Help1Lemma
Formulate reusable predicates for pointwise refinements for a single
variable and for a set of variables and a predicate
(TerminationRefinement) to express that one statement is at least as
terminating as the other.

The original formulation of the
Corollary_5_2HelpToCorollary_5_4Help1Lemma seems to be incorrect,
wrongly using the same state for the pointwise "val" and the everywhere
[]. After completing the fix at its call site, the remaining "assume" in
the re-written proof should be removed.